+++ packages/circuit-ui/components/FieldAtoms/FieldValidationHint.tsx
@@ -155,6 +155,7 @@ export const FieldValidationHint = ({
 
   return (
     <Wrapper
+      {...props}
       invalid={props.invalid}
       showValid={props.showValid}
       hasWarning={props.hasWarning}