+++ packages/circuit-ui/components/SidePanel/components/DesktopSidePanel/DesktopSidePanel.spec.tsx
@@ -71,7 +71,7 @@ describe('DesktopSidePanel', () => {
   });
 
   /**
-   * FIXME: calling axe here can trigger an act() warning.
+   * FIXME: calling axe here triggers an act() warning.
    */
   it('should meet accessibility guidelines', async () => {
     jest.useRealTimers();